Remove old reader files
