+++ src/Spout/Reader/AbstractReader.php
@@ -4,7 +4,6 @@ namespace Box\Spout\Reader;
 
 use Box\Spout\Common\Exception\IOException;
 use Box\Spout\Reader\Exception\ReaderNotOpenedException;
-use Box\Spout\Reader\Exception\EndOfFileReachedException;
 
 /**
  * Class AbstractReader
@@ -14,17 +13,8 @@ use Box\Spout\Reader\Exception\EndOfFileReachedException;
  */
 abstract class AbstractReader implements ReaderInterface
 {
-    /** @var int Used to keep track of the row index */
-    protected $currentRowIndex = 0;
-
     /** @var bool Indicates whether the stream is currently open */
     protected $isStreamOpened = false;
-
-    /** @var bool Indicates whether all rows have been read */
-    protected $hasReachedEndOfFile = false;
-
-    /** @var array Buffer used to store the row data, while checking if there are more rows to read */
-    protected $rowDataBuffer = null;
 
     /** @var \Box\Spout\Common\Helper\GlobalFunctionsHelper Helper to work with global functions */
     protected $globalFunctionsHelper;
@@ -38,11 +28,11 @@ abstract class AbstractReader implements ReaderInterface
     abstract protected function openReader($filePath);
 
     /**
-     * Reads and returns next row if available.
+     * Returns an iterator to iterate over sheets.
      *
-     * @return array|null Array that contains the data for the read row or null at the end of the file
+     * @return \Iterator To iterate over sheets
      */
-    abstract protected function read();
+    abstract public function getConcreteSheetIterator();
 
     /**
      * Closes the reader. To be used after reading the file.
@@ -80,9 +70,6 @@ abstract class AbstractReader implements ReaderInterface
             }
         }
 
-        $this->currentRowIndex = 0;
-        $this->hasReachedEndOfFile = false;
-
         try {
             $this->openReader($filePath);
             $this->isStreamOpened = true;
@@ -103,82 +90,18 @@ abstract class AbstractReader implements ReaderInterface
     }
 
     /**
-     * Returns whether all rows have been read (i.e. if we are at the end of the file).
-     * To know if the end of file has been reached, it uses a buffer. If the buffer is
-     * empty (meaning, nothing has been read or previous read line has been consumed), then
-     * it reads the next line, store it in the buffer for the next time or flip a variable if
-     * the end of file has been reached.
+     * Returns an iterator to iterate over sheets.
      *
-     * @return bool Whether all rows have been read (i.e. if we are at the end of the file)
-     * @throws \Box\Spout\Reader\Exception\ReaderNotOpenedException If the stream was not opened first
+     * @return \Iterator To iterate over sheets
+     * @throws \Box\Spout\Reader\Exception\ReaderNotOpenedException If called before opening the reader
      */
-    public function hasNextRow()
+    public function getSheetIterator()
     {
         if (!$this->isStreamOpened) {
-            throw new ReaderNotOpenedException('Stream should be opened first.');
+            throw new ReaderNotOpenedException('Reader should be opened first.');
         }
 
-        if ($this->hasReachedEndOfFile) {
-            return false;
-        }
-
-        // if the buffer contains unprocessed row
-        if (!$this->isRowDataBufferEmpty()) {
-            return true;
-        }
-
-        // otherwise, try to read the next line line, and store it in the buffer
-        $this->rowDataBuffer = $this->read();
-
-        // if the buffer is still empty after reading a row, it means end of file was reached
-        $this->hasReachedEndOfFile = $this->isRowDataBufferEmpty();
-
-        return (!$this->hasReachedEndOfFile);
-    }
-
-    /**
-     * Returns next row if available. The row is either retrieved from the buffer if it is not empty or fetched by
-     * actually reading the file.
-     *
-     * @return array Array that contains the data for the read row
-     * @throws \Box\Spout\Common\Exception\IOException If the stream was not opened first
-     * @throws \Box\Spout\Reader\Exception\EndOfFileReachedException
-     */
-    public function nextRow()
-    {
-        if (!$this->hasNextRow()) {
-            throw new EndOfFileReachedException('End of file was reached. Cannot read more rows.');
-        }
-
-        // Get data from buffer (if the buffer was empty, it was filled by the call to hasNextRow())
-        $rowData = $this->rowDataBuffer;
-
-        // empty buffer to mark the row as consumed
-        $this->emptyRowDataBuffer();
-
-        $this->currentRowIndex++;
-
-        return $rowData;
-    }
-
-    /**
-     * Returns whether the buffer where the row data is stored is empty
-     *
-     * @return bool
-     */
-    protected function isRowDataBufferEmpty()
-    {
-        return ($this->rowDataBuffer === null);
-    }
-
-    /**
-     * Empty the buffer that stores row data
-     *
-     * @return void
-     */
-    protected function emptyRowDataBuffer()
-    {
-        $this->rowDataBuffer = null;
+        return $this->getConcreteSheetIterator();
     }
 
     /**
@@ -190,6 +113,12 @@ abstract class AbstractReader implements ReaderInterface
     {
         if ($this->isStreamOpened) {
             $this->closeReader();
+
+            $sheetIterator = $this->getConcreteSheetIterator();
+            if ($sheetIterator) {
+                $sheetIterator->end();
+            }
+
             $this->isStreamOpened = false;
         }
     }
