--- src/Spout/Reader/Exception/EndOfFileReachedException.php
@@ -1,12 +1,0 @@
-<?php
-
-namespace Box\Spout\Reader\Exception;
-
-/**
- * Class EndOfFileReachedException
- *
- * @package Box\Spout\Reader\Exception
- */
-class EndOfFileReachedException extends ReaderException
-{
-}
