--- src/Spout/Reader/Exception/EndOfWorksheetsReachedException.php
@@ -1,12 +1,0 @@
-<?php
-
-namespace Box\Spout\Reader\Exception;
-
-/**
- * Class EndOfWorksheetsReachedException
- *
- * @package Box\Spout\Reader\Exception
- */
-class EndOfWorksheetsReachedException extends ReaderException
-{
-}
