--- src/Spout/Reader/Exception/NoWorksheetsFoundException.php
@@ -1,12 +1,0 @@
-<?php
-
-namespace Box\Spout\Reader\Exception;
-
-/**
- * Class NoWorksheetsFoundException
- *
- * @package Box\Spout\Reader\Exception
- */
-class NoWorksheetsFoundException extends ReaderException
-{
-}
