+++ src/Spout/Reader/ReaderFactory.php
@@ -19,7 +19,7 @@ class ReaderFactory
      * This creates an instance of the appropriate reader, given the type of the file to be read
      *
      * @param  string $readerType Type of the reader to instantiate
-     * @return \Box\Spout\Reader\CSV|\Box\Spout\Reader\XLSX
+     * @return \Box\Spout\Reader\CSV\Reader|\Box\Spout\Reader\XLSX\Reader
      * @throws \Box\Spout\Common\Exception\UnsupportedTypeException
      */
     public static function create($readerType)
@@ -28,10 +28,10 @@ class ReaderFactory
 
         switch ($readerType) {
             case Type::CSV:
-                $reader = new CSV();
+                $reader = new CSV\Reader();
                 break;
             case Type::XLSX:
-                $reader = new XLSX();
+                $reader = new XLSX\Reader();
                 break;
             default:
                 throw new UnsupportedTypeException('No readers supporting the given type: ' . $readerType);
