--- src/Spout/Reader/ReaderFactory2.php
@@ -1,44 +1,0 @@
-<?php
-
-namespace Box\Spout\Reader;
-
-use Box\Spout\Common\Exception\UnsupportedTypeException;
-use Box\Spout\Common\Helper\GlobalFunctionsHelper;
-use Box\Spout\Common\Type;
-
-/**
- * Class ReaderFactory2
- * This factory is used to create readers, based on the type of the file to be read.
- * It supports CSV and XLSX formats.
- *
- * @package Box\Spout\Reader
- */
-class ReaderFactory2
-{
-    /**
-     * This creates an instance of the appropriate reader, given the type of the file to be read
-     *
-     * @param  string $readerType Type of the reader to instantiate
-     * @return \Box\Spout\Reader\CSV\Reader|\Box\Spout\Reader\XLSX\Reader
-     * @throws \Box\Spout\Common\Exception\UnsupportedTypeException
-     */
-    public static function create($readerType)
-    {
-        $reader = null;
-
-        switch ($readerType) {
-            case Type::CSV:
-                $reader = new CSV\Reader();
-                break;
-            case Type::XLSX:
-                $reader = new XLSX\Reader();
-                break;
-            default:
-                throw new UnsupportedTypeException('No readers supporting the given type: ' . $readerType);
-        }
-
-        $reader->setGlobalFunctionsHelper(new GlobalFunctionsHelper());
-
-        return $reader;
-    }
-}
