+++ src/Spout/Reader/ReaderInterface.php
@@ -20,26 +20,12 @@ interface ReaderInterface
     public function open($filePath);
 
     /**
-     * Returns whether all rows have been read (i.e. if we are at the end of the file).
-     * To know if the end of file has been reached, it uses a buffer. If the buffer is
-     * empty (meaning, nothing has been read or previous read line has been consumed), then
-     * it reads the next line, store it in the buffer for the next time or flip a variable if
-     * the end of file has been reached.
+     * Returns an iterator to iterate over sheets.
      *
-     * @return bool
-     * @throws \Box\Spout\Common\Exception\IOException If the stream was not opened first
+     * @return \Iterator To iterate over sheets
+     * @throws \Box\Spout\Reader\Exception\ReaderNotOpenedException If called before opening the reader
      */
-    public function hasNextRow();
-
-    /**
-     * Returns next row if available. The row is either retrieved from the buffer if it is not empty or fetched by
-     * actually reading the file.
-     *
-     * @return array Array that contains the data for the read row
-     * @throws \Box\Spout\Common\Exception\IOException If the stream was not opened first
-     * @throws \Box\Spout\Reader\Exception\EndOfFileReachedException
-     */
-    public function nextRow();
+    public function getSheetIterator();
 
     /**
      * Closes the reader, preventing any additional reading
